Link unit tests against mudlet_core library
Instead of each test compiling source files directly, link against the
mudlet_core static library. This avoids redundant compilation and
ensures tests exercise the exact same code that ships.

- Extract HyperlinkStyling to own header to break TBuffer/TLinkStore
  circular dependency
- Remove LinkStore_Test conditional compilation guards
- Add sanitizer support and ASAN leak suppression for unit tests
